Refactored packages to fit naming
